context and slider made function to reuse
