+++ LiFrame/Manager+Extenstion/UIView+Extension.swift
@@ -39,56 +39,4 @@ extension UIView {
                })
            }
        }
-//    // Border Color
-//    @IBInspectable var lkBorderColor: UIColor? {
-//        get {
-//            guard let borderColor = layer.borderColor else {
-//                return nil
-//            }
-//            return UIColor(cgColor: borderColor)
-//        }
-//        set {
-//            layer.borderColor = newValue?.cgColor
-//        }
-//    }
-//
-//    // Border width
-//    @IBInspectable var lkBorderWidth: CGFloat {
-//        get {
-//            return layer.borderWidth
-//        }
-//        set {
-//            layer.borderWidth = newValue
-//        }
-//    }
-//
-//    // Corner radius
-//    @IBInspectable var lkCornerRadius: CGFloat {
-//        get {
-//            return layer.cornerRadius
-//        }
-//        set {
-//            layer.cornerRadius = newValue
-//        }
-//    }
-//
-//    func stickSubView(_ objectView: UIView) {
-//        objectView.removeFromSuperview()
-//        addSubview(objectView)
-//        objectView.translatesAutoresizingMaskIntoConstraints = false
-//        objectView.topAnchor.constraint(equalTo: topAnchor).isActive = true
-//        objectView.leadingAnchor.constraint(equalTo: leadingAnchor).isActive = true
-//        objectView.trailingAnchor.constraint(equalTo: trailingAnchor).isActive = true
-//        objectView.bottomAnchor.constraint(equalTo: bottomAnchor).isActive = true
-//    }
-//
-//    func stickSubView(_ objectView: UIView, inset: UIEdgeInsets) {
-//        objectView.removeFromSuperview()
-//        addSubview(objectView)
-//        objectView.translatesAutoresizingMaskIntoConstraints = false
-//        objectView.topAnchor.constraint(equalTo: topAnchor, constant: inset.top).isActive = true
-//        objectView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: inset.left).isActive = true
-//        objectView.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -inset.right).isActive = true
-//        objectView.bottomAnchor.constraint(equalTo: bottomAnchor, constant: -inset.bottom).isActive = true
-//    }
 }
